gpt_analyzer/requirements.txt
openai>=1.0.0
pandas
python-dotenv